Refactor #11 - 메소드 방식 변경
* util → @Component 방식으로 변경
