+++ build.gradle
@@ -16,6 +16,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-devtools'
+    implementation 'org.jsoup:jsoup:1.14.3'
     implementation 'org.springdoc:springdoc-openapi-ui:1.6.6'
 
     compileOnly 'org.projectlombok:lombok'
