+++ src/main/java/study/subject/upa/global/util/Constants.java
@@ -1,0 +1,17 @@
+package study.subject.upa.global.util;
+
+/**
+ * @Class : Constants
+ * @Author : KDW
+ * @Date : 2022-03-14
+ * @Description : 상수 패턴 Class
+ */
+public class Constants {
+
+    public static final String HTML_PATTERN = "<(/)?([a-zA-Z]*)(\\s[a-zA-Z]*=[^>]*)?(\\s)*(/)?>";
+    public static final String NOT_ALPHA_NUMERIC_PATTERN = "[\\W_]";
+    public static final String NUM_PATTERN = "[0-9]";
+    public static final String ENG_PATTERN = "[a-zA-Z]";
+    public static final String WHITESPACE_PATTERN = "\\s";
+    public static final String EMPTY = "";
+}
